+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixes
+- `execute`: cancel the Redash job when it times out or is interrupted (Ctrl-C), instead of leaving it running server-side with nobody waiting for the result ([#21](https://github.com/mozilla/stmo-cli/issues/21))
+
 ## [0.9.0] - 2026-07-16
 
 ### Features